+++ src/errors/NotFoundException.error.ts
@@ -1,4 +1,4 @@
-import { CustomException } from "./CustomException";
+import { CustomException } from "./CustomException.error";
 
 export class NotFoundException extends CustomException {
   statusCode = 404;
